Make `MagicIdx` a 32 bit value, not a 16 bit one
Allows for mroe than 2^16 magic values

+++ crates/vm/src/instructions/instruction.rs
@@ -4,7 +4,7 @@ pub type RegIdx = u8;
 pub type ConstIdx = u16;
 pub type HeapIdx = u8;
 pub type ProtoIdx = u8;
-pub type MagicIdx = u16;
+pub type MagicIdx = u32;
 pub type InstIdx = u32;
 
 macro_rules! for_each_instruction {
